+++ eslint-config/CHANGELOG.md
@@ -6,15 +6,30 @@ All notable changes to this project will be documented in this file. The format 
 
 ---
 
-## [Unreleased](https://github.com/greylocklabs/js/compare/1.2.0...HEAD)
+## [Unreleased](https://github.com/greylocklabs/js/compare/1.4.0...HEAD)
 
 - N/A
+
+## [1.4.0](https://github.com/greylocklabs/js/compare/1.3.0...1.4.0) - 2017-11-15
+
+### Changed
+
+- [`no-param-reassign`](https://eslint.org/docs/rules/no-param-reassign) - Allow modifications to the properties of function arguments
+- [`arrow-parens`](https://eslint.org/docs/rules/arrow-parens) - Always use () around arrow function params
+- [`object-shorthand`](https://eslint.org/docs/rules/object-shorthand) - Relaxed rule to `consistent-as-needed` setting
+- [`func-style`](https://eslint.org/docs/rules/func-style) - Allow for arrow functions
+- [`eslint-plugin-flowtype`](https://github.com/gajus/eslint-plugin-flowtype) changes:
+    - [`flowtype/require-parameter-type`](https://github.com/gajus/eslint-plugin-flowtype#eslint-plugin-flowtype-rules-require-parameter-type) - now throws error
+    - [`flowtype/require-return-type`](https://github.com/gajus/eslint-plugin-flowtype#eslint-plugin-flowtype-rules-require-return-type) - now throws error
+    - [`flowtype/require-valid-file-annotation`](https://github.com/gajus/eslint-plugin-flowtype#require-valid-file-annotation) - now throws error
+- [`eslint-plugin-security`](https://github.com/nodesecurity/eslint-plugin-security) - Allow for object bracket notation
 
 ## [1.3.0](https://github.com/greylocklabs/js/compare/1.2.0...1.3.0) - 2017-11-13
 
 ### Added
 
 - Support for Flow via [eslint-plugin-flowtype](https://github.com/gajus/eslint-plugin-flowtype)
+    - Optional for projects that don't use flow - only checks files with flow annotation
 
 ## [1.2.0](https://github.com/greylocklabs/js/compare/1.1.0...1.2.0) - 2017-11-13